Completes (Top Moving Ingredients)
1. Creates the query to provide the necessary data
from the db to the endpoint. Relates to TA-12.

2. Creates the endpoint 'get/top-moving-ingredients'
to serve the data to the client.
Small change to the query to name the column from
ingredient_name  to name.
Relates to TA-12.

3. Creates the client side query to fetch the top
moving ingredients data from the server.
Relates to TA-12.

4. Implement the "real" data in the Top Moving
Ingredients Pie chart and makes some changes to
better display the data by creating a custom
JSX element for the chart labels. Closes TA-12.